packages/hearth/requirements.txt
# Hearth Requirements

# API clients
httpx>=0.25.0
anthropic>=0.40.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Telegram
python-telegram-bot>=20.0

# Scheduling
schedule>=1.2.0

# CLI
click>=8.1.0
rich>=13.0

# Database
aiosqlite>=0.20.0

# Testing
pytest>=7.0
pytest-asyncio>=0.21
